move libraries to the common area
