+++ cpp/library/TreeStructures.hh
